+++ webapp/components/hooks/useIsAssessmentLocked.js
